requirements.txt
pymongo>=4.6
pandas>=2.0
numpy>=1.26
scikit-learn>=1.4
plotly>=5.20
kaleido>=0.2.1
python-dotenv>=1.0
requests>=2.31
scipy>=1.12
fastapi>=0.111
uvicorn[standard]>=0.29
